Three root causes from the parallel drawing round. Nine of its fifteen documents
screened out first — two as the version gap and four where the catalogued cause was
not reproducible against 26.2 at all (the logo, the photograph and the background
wash all present and correctly placed).

* a floating VML shape's `z-index` was never read, so every one painted in document
  order in front of the text; the sign is the layer (`DomainMapper_Impl.cxx`:5157)
  and a `z-index` outranks every `relativeHeight` whatever the numbers
  (`GraphicHelpers.cxx`:279-330);
* a row whose cells all hide their mark is exactly `w:trHeight` tall
  (`DomainMapperTableHandler.cxx`:1157-1162), with tdf#77417's trailing-space trim
  deciding which cells count as empty;
* a picture inside an `xdr:grpSp` on a sheet was read for its bounds and never
  drawn (`drawingfragment.cxx`:198).

+++ dotnet/src/Paperless.WordProcessing/Ooxml/DocxFrames.cs
@@ -196,12 +196,12 @@ internal static class DocxFrames
     /// range round-trips rather than most of it. An unparseable or absent value is zero, which sorts
     /// below every anchor that declares one and leaves document order untouched.
     /// </remarks>
-    private static uint ZOrder(XElement? anchor) =>
+    private static long ZOrder(XElement? anchor) =>
         anchor?.Attribute("relativeHeight")?.Value is { } text
         && uint.TryParse(text, System.Globalization.NumberStyles.Integer,
                          System.Globalization.CultureInfo.InvariantCulture, out uint z)
             ? z
-            : 0u;
+            : 0L;
 
     /// <summary>Whether the anchor asks for one of the four wraps that leave a hole in the text.</summary>
     private static bool WrapsAside(XElement anchor)
